Add new Russian/Latin combined keyboard layout.
This will be very helpful for people that uses both Russian and US English keyboard layouts on Windows and switch between them (most Russian users in the world).
For this case using only US English guacd RDP keymap or even creating and using some Russian keymap will not be sufficient because of users layout switching that can't be detected by Guacamole (at least for now). The problem is that when US English guacd RDP keymap or failsafe one is used, it will do the work for english (latin) typing using US English Windows keyboard layout only. Switching to Russian one will generate russian symbols not as RDP scancodes but as RDP unicode events that Windows RDP host will threat as special case and will issue VK_PACKET instead of regular keycodes. This will lead to great pain, because almost all apps will not handle VK_PACKET events properly. For example, shortcut key combinations like Ctrl+C and Ctrl+V will not work in most cases if user will switch to Russian Windows layout. Also there will be a problem with entering special symbols like ":", ";", ",", ".", etc. Using Failsafe keymap will "fix" special symbols but also will completely break shortcut combinations, even when user types in US English Windows layout.

The effort of creating some "combined" keymap is directed to overcome these issues. We know that Guacamole Client sends key symbols to guacd and keymap is used to convert key symbols to RDP scancodes. Keymap file consists of two main parts: left and right. On the left there are RDP scancodes written with any additional modifiers that must or must not be "pressed" to generate a valid key event with this scancode. On the right part there are key symbols written as they sent by Guacamole Client. So the keymap subsystem translates what is received (from the right part) to what to send (to the left part accordingly).
So the main task is to create a union keymap where all different key symbols from Russian and English keyboard layout will be combined and all the "same" symbols that are placed on different keys will be removed (these symbols will be issued as unicode VK_PACKED events). This way will eliminate most pain related to the issue: almost all the symbols will be translated to needed scancodes (so all the functionality including shortcuts will work in all keyboard layouts), all the symbols from two layouts will be on the right places and only a small subset of symbols (that is same by nature but are placed on different keys in different layouts) will be issued as unicode. This is also the main drawback of this decision, becuase user must ensure that layouts on RDP client and host are in sync. So at least the same key combination to switch layouts on client and on host must be used and change layout by hand on either client or host if they are out of sync.

I do not know if the other workaround can be applied to fix this issues, but at least this fix is better than nothing.